% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_deps.R
\name{proj_check_deps}
\alias{proj_check_deps}
\title{Check dependencies}
\usage{
proj_check_deps(path = usethis::proj_get())
}
\arguments{
\item{path}{\code{character}, path to the project-root directory.}
}
\value{
\code{list} with elements \code{missing} and \code{extra},
containing names of missing and extra package dependencies
}
\description{
This uses \code{\link[renv:dependencies]{renv::dependencies()}}, which scans your project directory for
package dependencies. It compares those detected in the code with those
declared in the \code{DESCRIPTION} file to compile and return a list of
missing and extra packages.
}
\examples{
if (FALSE) {
  # not run because it depends on side-effects
  proj_check_deps()
}

}
