% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_create.R
\name{proj_create}
\alias{proj_create}
\title{Create project}
\usage{
proj_create(
  path,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive(),
  fields = list()
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{rstudio}{If \code{TRUE}, calls \code{\link[usethis:use_rstudio]{use_rstudio()}} to make the new package or
project into an \href{https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects}{RStudio Project}.
If \code{FALSE} and a non-package project, a sentinel \code{.here} file is placed so
that the directory can be recognized as a project by the
\href{https://here.r-lib.org}{here} or
\href{https://rprojroot.r-lib.org}{rprojroot} packages.}

\item{open}{If \code{TRUE}, \link[usethis:proj_activate]{activates} the new project:
\itemize{
\item If RStudio desktop, the package is opened in a new session.
\item If on RStudio server, the current RStudio project is activated.
\item Otherwise, the working directory and active project is changed.
}}

\item{fields}{A named list of fields to add to \code{DESCRIPTION}, potentially
overriding default values. See \code{\link[usethis:use_description]{use_description()}} for how you can set
personalized defaults using package options}
}
\value{
Invisible \code{NULL}, called for side effects
}
\description{
To create a new project, this function calls \code{\link[usethis:create_package]{usethis::create_project()}},
then \code{\link[=proj_use_description]{proj_use_description()}} to create \code{DESCRIPTION} file.
}
\examples{
# not run because it evokes side effects
if (FALSE) {
  proj_create("../new-project-name")
}
}
