% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{proj_create_dir_target}
\alias{proj_create_dir_target}
\title{Create target directory}
\usage{
proj_create_dir_target(name)
}
\arguments{
\item{name}{\code{character} name of the workflow component.}
}
\value{
Invisible NULL, called for side effects.
}
\description{
Create target directory
}
\examples{
# not run because it creates side effects
\dontrun{
  proj_create_dir_target("01-clean")
}
}
