% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{proj_dir_info}
\alias{proj_dir_info}
\title{Local directory information}
\usage{
proj_dir_info(
  path = ".",
  tz = "UTC",
  cols = c("path", "type", "size", "modification_time"),
  ...
)
}
\arguments{
\item{path}{\code{character} path for directory listing.}

\item{tz}{\code{character} Olson time-zone to use for datetimes;
specify \code{NULL} to use system time-zone.}

\item{cols}{\code{character} vector of column-names to return from
\code{\link[fs:dir_ls]{fs::dir_info()}}; specify \code{NULL} to return all columns.}

\item{...}{additional arguments passed to \code{\link[fs:dir_ls]{fs::dir_info()}}.}
}
\value{
\code{data.frame} with S3 classes \code{"tbl_df"} and \code{"tbl"}, aka a "tibble".
}
\description{
This provides a different take on \code{\link[fs:dir_ls]{fs::dir_info()}}, by default
returning only a (hopefully useful) subset of the information.
}
\examples{
  proj_dir_info()
}
