% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_deps.R
\name{proj_install_deps}
\alias{proj_install_deps}
\title{Install dependencies}
\usage{
proj_install_deps(path = usethis::proj_get(), dependencies = TRUE, ...)
}
\arguments{
\item{path}{\code{character} path to project root.}

\item{dependencies}{Which dependencies do you want to check?
Can be a character vector (selecting from "Depends", "Imports",
"LinkingTo", "Suggests", or "Enhances"), or a logical vector.

\code{TRUE} is shorthand for "Depends", "Imports", "LinkingTo" and
"Suggests". \code{NA} is shorthand for "Depends", "Imports" and "LinkingTo"
and is the default. \code{FALSE} is shorthand for no dependencies (i.e.
just check this package, not its dependencies).

The value "soft" means the same as \code{TRUE}, "hard" means the same as \code{NA}.

You can also specify dependencies from one or more additional fields,
common ones include:
\itemize{
\item Config/Needs/website - for dependencies used in building the pkgdown site.
\item Config/Needs/coverage for dependencies used in calculating test coverage.
}}

\item{...}{other arguments passed to \code{\link[remotes:install_deps]{remotes::install_deps()}}.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
Use to install the project's package dependencies.
This is a thin wrapper to \code{\link[remotes:install_deps]{remotes::install_deps()}}; by default, it installs
all "Depends", "Imports", "Suggests", and "LinkingTo".
}
\examples{
# not run because it produces side effects
if (FALSE) {
  proj_install_deps()
}
}
