% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{proj_path_target}
\alias{proj_path_target}
\alias{proj_path_source}
\title{Create function to access directory}
\usage{
proj_path_target(name)

proj_path_source(name)
}
\arguments{
\item{name}{\code{character} name of the workflow component.}
}
\value{
\code{function} that acts like \code{\link[here:here]{here::here()}},
returning a \code{character} path.
}
\description{
Create function to access directory
}
\examples{
  proj_path_target("01-clean")
  proj_path_source("01-clean")
}
