% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{proj_refresh_deps}
\alias{proj_refresh_deps}
\title{Refresh dependencies}
\usage{
proj_refresh_deps(
  path = usethis::proj_get(),
  remove_extra = FALSE,
  dependencies = TRUE,
  ...
)
}
\arguments{
\item{path}{\code{character}, path to the project directory. If your current
working-directory in is in the project, the default will do the right thing.}

\item{remove_extra}{\code{logical}, indicates to remove dependency-declarations
that \code{\link[renv:dependencies]{renv::dependencies()}} can't find being used.}

\item{dependencies}{Which dependencies do you want to check?
Can be a character vector (selecting from "Depends", "Imports",
"LinkingTo", "Suggests", or "Enhances"), or a logical vector.

\code{TRUE} is shorthand for "Depends", "Imports", "LinkingTo" and
"Suggests". \code{NA} is shorthand for "Depends", "Imports" and "LinkingTo"
and is the default. \code{FALSE} is shorthand for no dependencies (i.e.
just check this package, not its dependencies).

The value "soft" means the same as \code{TRUE}, "hard" means the same as \code{NA}.

You can also specify dependencies from one or more additional fields,
common ones include:
\itemize{
\item Config/Needs/website - for dependencies used in building the pkgdown site.
\item Config/Needs/coverage for dependencies used in calculating test coverage.
}}

\item{...}{other arguments passed to \code{\link[remotes:install_deps]{remotes::install_deps()}}.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
This calls \code{proj_update_deps()}, then \code{proj_install_deps()}; use to refresh
your package dependencies.
}
\examples{
# not run because it produces side effects
if (FALSE) {
  proj_refresh_deps()
}
}
