% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{proj_rmd_render}
\alias{proj_rmd_render}
\title{Render in new R session}
\usage{
proj_rmd_render(input, ...)
}
\arguments{
\item{input}{\code{character}, path to the file to be rendered.}

\item{...}{other args passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
The goal of this function is to mimic the behavior of the "knit" button in
the RStudio IDE - it will start a new R session in which the \code{.Rmd} file
is rendered, using \code{\link[rmarkdown:render]{rmarkdown::render()}}. By starting a new session,
we avoid package namespaces from being attached from previous renderings.
}
\examples{
\dontrun{
  # not run because it depends on and creates side effects
  proj_rmd_render("00-input.Rmd")
}
}
