% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_description.R
\name{proj_use_description}
\alias{proj_use_description}
\title{Create DESCRIPTION file}
\usage{
proj_use_description(fields = list(), update_deps = TRUE)
}
\arguments{
\item{fields}{A named list of fields to add to \code{DESCRIPTION}, potentially
overriding default values. See \code{\link[usethis:use_description]{use_description()}} for how you can set
personalized defaults using package options}

\item{update_deps}{\code{logical} indicates to update dependencies in
DESCRIPTION file}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
To create a DESCRIPTION file for your existing project, this function calls
\code{\link[usethis:use_description]{usethis::use_description()}} and, optionally, \code{\link[=proj_update_deps]{proj_update_deps()}} to
update the declared dependencies.
}
\examples{
# not run because it produces side effects
if (FALSE) {
  proj_use_description()
}
}
