% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{proj_workflow_config}
\alias{proj_workflow_config}
\title{Get workflow configuration}
\usage{
proj_workflow_config(path_proj)
}
\arguments{
\item{path_proj}{\code{character} path to workflow directory,
relative to the project directory.}
}
\value{
\code{NULL}, or \code{list} describing workflow configuration
}
\description{
Looks for a file named \verb{_projthis.yml} in \code{path_proj}. If present, reads
using \code{\link[yaml:read_yaml]{yaml::read_yaml()}}; if not present, returns \code{NULL}.
}
\keyword{internal}
