% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{proj_workflow_render}
\alias{proj_workflow_render}
\title{Render workflow}
\usage{
proj_workflow_render(
  path_proj = "workflow",
  output_options = list(html_preview = FALSE),
  ...
)
}
\arguments{
\item{path_proj}{\code{character} path to workflow directory,
relative to the project directory.}

\item{output_options}{\code{list} of output options that can override the
RMarkdown file metadata. The default sets \code{html_preview = FALSE} to avoid
HTML files being created for \code{github_document}.}

\item{...}{other arguments passed on to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
This renders each of the \code{.Rmd} files in the workflow in alphabetical
order, with \code{README.Rmd} last. This order is important because it preserves
the direction of the data dependencies. Each \code{.Rmd} file is rendered in its
own R session, using \code{\link[=proj_rmd_render]{proj_rmd_render()}}
}
\examples{
\dontrun{
  # not run because it creates side effects
  proj_workflow_render()
}
}
