% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{proj_workflow_use_action}
\alias{proj_workflow_use_action}
\title{Use GitHub Action}
\usage{
proj_workflow_use_action(
  name = "project-workflow",
  save_as = NULL,
  ignore = TRUE,
  open = TRUE
)
}
\arguments{
\item{name}{For \code{use_github_action()}: Name of one of the example workflows
from \url{https://github.com/r-lib/actions/tree/v2/examples}. Examples:
"pkgdown", "check-standard.yaml".

For \code{use_github_actions_badge()}: Name of the workflow's YAML configuration
file. Examples: "R-CMD-check", "R-CMD-check.yaml".

If \code{name} has no extension, we assume it's \code{.yaml}.}

\item{save_as}{Name of the local workflow file. Defaults to \code{name} or
\code{fs::path_file(url)} for \code{use_github_action()}. Do not specify any other
part of the path; the parent directory will always be \code{.github/workflows},
within the active project.}

\item{ignore}{Should the newly created file be added to \code{.Rbuildignore}?}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
You can use a GitHub Action to automate the building and deployment of your
project. This function wraps \code{\link[usethis:github_actions]{usethis::use_github_action()}}  to install an
Actions file that will:
\itemize{
\item install R
\item install and cache your dependencies (using DESCRIPTION)
}

You will need to adapt the sections in your Actions file to:
\itemize{
\item build your project
\item deploy your project
}

To help you draw the
\href{https://knowyourmeme.com/memes/how-to-draw-an-owl}{"rest of the owl"}, the
\href{https://github.com/r-lib/actions}{r-lib actions repository} is a great place
to start.
}
\examples{
# not run because it produces side effects
if (FALSE) {
  proj_workflow_use_action()
}
}
