% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{proj_workflow_use_qmd}
\alias{proj_workflow_use_qmd}
\title{Use opinionated qmd template}
\usage{
proj_workflow_use_qmd(
  name,
  path_proj = NULL,
  open = rlang::is_interactive(),
  ignore = FALSE
)
}
\arguments{
\item{name}{\code{character} name of the workflow component.}

\item{path_proj}{\code{character} path to workflow directory,
relative to the project directory.}

\item{open}{\code{logical} indicates to open the file for interactive editing.}

\item{ignore}{\code{logical} indicates to add this file to \code{.Rbuildignore}.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
Use opinionated qmd template
}
\examples{
# not run because it creates side effects
\dontrun{
  # creates file `01-clean.qmd`
  proj_workflow_use_rmd("01-clean")
}
}
