% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{proj_workflow_use_rmd}
\alias{proj_workflow_use_rmd}
\title{Use opinionated Rmd template}
\usage{
proj_workflow_use_rmd(
  name,
  path = NULL,
  open = rlang::is_interactive(),
  ignore = FALSE
)
}
\arguments{
\item{name}{\code{character} name of the file, extension not necessary.}

\item{path}{\code{character} path to workflow directory,
relative to project root.}

\item{open}{\code{logical} indicates to open the file for interactive editing.}

\item{ignore}{\code{logical} indicates to add this file to \code{.Rbuildignore}.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
This uses \code{\link[usethis:use_template]{usethis::use_template()}} to create a new, templated RMarkdown file
in a flow directory.
}
\examples{
\dontrun{
  # creates file `01-clean.Rmd`
  proj_workflow_use_rmd("01-clean")
}
}
