% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{proj_workflow_render}
\alias{proj_workflow_render}
\title{Render workflow}
\usage{
proj_workflow_render(path = ".", envir = new.env(), ...)
}
\arguments{
\item{path}{\code{character} path to workflow directory,
relative to project root.}

\item{envir}{\code{environment} in which code chunks are to be evaluated.}

\item{...}{other arguments passed on to \code{\link[rmarkdown:render_site]{rmarkdown::render_site()}}.}
}
\value{
\code{character} name of the site output file,
relative to the input directory. Called principally for side effects.
}
\description{
Render workflow
}
\examples{
\dontrun{
  # not run because it depends on and produces side effects
  proj_workflow_render()
}
}
