% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{proj_workflow_render}
\alias{proj_workflow_render}
\title{Render workflow}
\usage{
proj_workflow_render(
  path_proj = "workflow",
  output_options = list(html_preview = FALSE),
  ...
)
}
\arguments{
\item{path_proj}{\code{character} path to workflow directory,
relative to the project directory.}

\item{output_options}{\code{list} of output options that can override the
RMarkdown file metadata. The default sets \code{html_preview = FALSE} to avoid
HTML files being created for \code{github_document}.}

\item{...}{other arguments passed on to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
In the absence of a \verb{_projthis.yml} file, this renders each of the \code{.Rmd}
files in the workflow in alphabetical order, with \code{README.Rmd} last. This
order is important because it preserves the direction of the data
dependencies.
}
\details{
A \verb{_projthis.yml} file might look something like this:\preformatted{render:
  first:
    00-import.Rmd
  last:
    99-publish.Rmd
}

If the workflow directory has a \verb{_projthis.yml} file:
\itemize{
\item Entries in \code{render$first} are rendered first.
\item Entries in \code{render$last} are rendered last, \strong{but}
\code{README.Rmd} is rendered very last.
\item If \code{README.Rmd} is specified in \code{render$first} or \code{render$last},
it is rendered there rather than very last.
\item All unspecified files are rendered in alphabetical order after \code{first},
and before \code{last}.
}

Each \code{.Rmd} file is rendered in its own R session,  using \code{\link[=proj_rmd_render]{proj_rmd_render()}}.
}
\examples{
\dontrun{
  # not run because it creates side effects
  proj_workflow_render()
}
}
