% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{proj_workflow_render}
\alias{proj_workflow_render}
\title{Render workflow}
\usage{
proj_workflow_render(
  path = "workflow",
  envir = new.env(),
  output_options = list(html_preview = FALSE),
  ...
)
}
\arguments{
\item{path}{\code{character} path to workflow directory,
relative to project root.}

\item{envir}{\code{environment} in which code chunks are to be evaluated.}

\item{output_options}{\code{list} of output options that can override the
RMarkdown file metadata.}

\item{...}{other arguments passed on to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
This renders each of the RMarkdown files in the workflow in alphabetical
order. This order is important because it preserves the direction of the
data dependencies.
}
\examples{
\dontrun{
  # not run because it creates side effects
  proj_workflow_render()
}
}
